+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 
 ## 0.2.0
 
+- Provide command `Start camel Application with JBang and debug`. It allows a one-click start and Camel debug in simple cases.
 - Provide error message and remind requirements in case of failure when trying to attach the debugger
 - Provide configuration snippets for Camel debugger launch configuration
 - Provide configuration snippets to launch Camel application ready to accept a Camel debugger connection using jbang or Maven